Split GitHub client from trait
